+++ boussole/cli/compile.py
@@ -3,8 +3,6 @@ import click
 import logging
 import os
 
-import six
-
 from boussole.compiler import SassCompileHelper
 from boussole.conf.discovery import Discover
 from boussole.conf.json_backend import SettingsBackendJson
@@ -28,7 +26,7 @@ def compile_command(context, backend, config):
     Compile Sass project sources to CSS
     """
     logger = logging.getLogger("boussole")
-    logger.info(u"Building project")
+    logger.info("Building project")
 
     # Discover settings file
     try:
@@ -43,16 +41,16 @@ def compile_command(context, backend, config):
         project = ProjectBase(backend_name=config_engine._kind_name)
         settings = project.backend_engine.load(filepath=config_filepath)
     except BoussoleBaseException as e:
-        logger.critical(six.text_type(e))
+        logger.critical(str(e))
         raise click.Abort()
 
-    logger.debug(u"Settings file: {} ({})".format(
+    logger.debug("Settings file: {} ({})".format(
                  config_filepath, config_engine._kind_name))
-    logger.debug(u"Project sources directory: {}".format(
+    logger.debug("Project sources directory: {}".format(
                  settings.SOURCES_PATH))
-    logger.debug(u"Project destination directory: {}".format(
+    logger.debug("Project destination directory: {}".format(
                  settings.TARGET_PATH))
-    logger.debug(u"Exclude patterns: {}".format(
+    logger.debug("Exclude patterns: {}".format(
                  settings.EXCLUDES))
 
     # Find all sources with their destination path
@@ -63,20 +61,20 @@ def compile_command(context, backend, config):
             excludes=settings.EXCLUDES
         )
     except BoussoleBaseException as e:
-        logger.error(six.text_type(e))
+        logger.error(str(e))
         raise click.Abort()
 
     # Build all compilable stylesheets
     compiler = SassCompileHelper()
     errors = 0
     for src, dst in compilable_files:
-        logger.debug(u"Compile: {}".format(src))
+        logger.debug("Compile: {}".format(src))
 
         output_opts = {}
         success, message = compiler.safe_compile(settings, src, dst)
 
         if success:
-            logger.info(u"Output: {}".format(message), **output_opts)
+            logger.info("Output: {}".format(message), **output_opts)
         else:
             errors += 1
             logger.error(message)