+++ boussole/cli/console_script.py
@@ -10,16 +10,16 @@ from boussole.cli.startproject import startproject_command
 from boussole.logs import init_logger
 
 
-# Help alias on '-h' argument
-CONTEXT_SETTINGS = dict(help_option_names=['-h', '--help'])
+# Help alias on "-h" argument
+CONTEXT_SETTINGS = dict(help_option_names=["-h", "--help"])
 
 # Default logger conf
-BOUSSOLE_LOGGER_CONF = ('DEBUG', 'INFO', 'WARNING', 'ERROR', 'CRITICAL', None)
+BOUSSOLE_LOGGER_CONF = ("DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL", None)
 
 
 @click.group(context_settings=CONTEXT_SETTINGS)
-@click.option('-v', '--verbose', type=click.IntRange(min=0, max=5), default=4,
-              metavar='INTEGER',
+@click.option("-v", "--verbose", type=click.IntRange(min=0, max=5), default=4,
+              metavar="INTEGER",
               help="An integer between 0 and 5, where '0' make a totaly "
               "silent output and '5' set level to DEBUG (the most verbose "
               "level). Default to '4' (Info level).")
@@ -44,8 +44,8 @@ def cli_frontend(ctx, verbose):
 
     # Init the default context that will be passed to commands
     ctx.obj = {
-        'verbosity': verbose,
-        'logger': root_logger,
+        "verbosity": verbose,
+        "logger": root_logger,
     }