+++ boussole/cli/startproject.py
@@ -3,33 +3,31 @@ import os
 import click
 import logging
 
-import six
-
 from boussole.exceptions import SettingsInvalidError
 from boussole.project import ProjectStarter
 
 
-@click.command('startproject', short_help="Create a new Sass project.")
-@click.option('--basedir', metavar='PATH',
+@click.command("startproject", short_help="Create a new Sass project.")
+@click.option("--basedir", metavar="PATH",
               prompt="Project base directory",
               help=("Base directory where settings filename and project "
                     "structure will be created."),
               default=".")
-@click.option('--sourcedir', metavar='PATH',
+@click.option("--sourcedir", metavar="PATH",
               prompt="Sources directory",
               help="Directory (within base dir) for your Sass sources.",
               default="scss")
-@click.option('--targetdir', metavar='PATH',
+@click.option("--targetdir", metavar="PATH",
               prompt="Target directory",
               help="Directory (within base dir) where to write compiled "
                    "files.",
               default="css")
-@click.option('--backend', metavar='STRING',
+@click.option("--backend", metavar="STRING",
               prompt="Settings format name",
-              type=click.Choice(['json', 'yaml']),
+              type=click.Choice(["json", "yaml"]),
               help="Settings format name",
               default="json")
-@click.option('--config', metavar='PATH',
+@click.option("--config", metavar="PATH",
               help="Settings file name",
               default=None)
 @click.pass_context
@@ -66,7 +64,7 @@ def startproject_command(context, basedir, sourcedir, targetdir,
             targetdir,
         ), cwd=os.getcwd())
     except SettingsInvalidError as e:
-        logger.critical(six.text_type(e))
+        logger.critical(str(e))
         raise click.Abort()
     else:
         logger.info(u"Project directory structure and configuration file have "
@@ -74,7 +72,7 @@ def startproject_command(context, basedir, sourcedir, targetdir,
 
         logger.info(u"Now you should start to create some Sass sources into "
                     "'{}', then compile them "
-                    "using:".format(results['sourcedir']))
+                    "using:".format(results["sourcedir"]))
 
         logger.info(u"    boussole compile "
-                    "--config={}".format(results['config']))
+                    "--config={}".format(results["config"]))