+++ docs/changelog.rst
@@ -5,6 +5,17 @@
 Changelog
 =========
 
+Version 2.0.0 - Unreleased
+--------------------------
+
+DONE:
+    * Drop Python2 support (remove six, unicode litteral, config);
+
+TODO:
+    * Drop pathtools dependency (only need in watcher);
+    * Update doc config, use livereload and with new RTD config file;
+    * Add new boussole config filename (boussole.json, etc..);
+
 Version 1.6.0 - 2021/02/04
 --------------------------